Pull SortingTableHead into ResultsTableBody
