+++ context/app/static/js/components/Search/ResultsTableBody/ResultsTableBody.jsx
@@ -1,14 +1,16 @@
 import React from 'react';
-
+import { SortingSelector } from 'searchkit';
 import { LightBlueLink } from 'js/shared-styles/Links';
 import { getByPath } from './utils';
-import { StyledTableBody, StyledTableRow, StyledTableCell } from './style';
+import { StyledTable, StyledTableBody, StyledTableRow, StyledTableCell } from './style';
+import SortingTableHead from '../SortingTableHead';
 
 function ResultsTableBody(props) {
-  const { hits, resultFields, detailsUrlPrefix, idField } = props;
+  const { hits, resultFields, detailsUrlPrefix, idField, sortOptions } = props;
   /* eslint-disable no-underscore-dangle, react/no-danger, jsx-a11y/control-has-associated-label */
   return (
-    <>
+    <StyledTable>
+      <SortingSelector options={sortOptions} listComponent={SortingTableHead} />
       {hits.map((hit) => (
         <StyledTableBody key={hit._id}>
           <StyledTableRow className={'highlight' in hit && 'before-highlight'}>
@@ -37,7 +39,7 @@ function ResultsTableBody(props) {
           )}
         </StyledTableBody>
       ))}
-    </>
+    </StyledTable>
   );
   /* eslint-enable no-underscore-dangle, react/no-danger, jsx-a11y/control-has-associated-label */
 }
