+++ context/app/static/js/components/Search/styleForTable.js
@@ -1,7 +1,6 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import ArrowUpward from '@material-ui/icons/ArrowUpwardRounded';
 import ArrowDownward from '@material-ui/icons/ArrowDownwardRounded';
-import Table from '@material-ui/core/Table';
 
 import { HeaderCell } from 'js/shared-styles/Table';
 
@@ -28,13 +27,4 @@ const StyledHeaderCell = styled(HeaderCell)`
   white-space: nowrap;
 `;
 
-const tableCss = css`
-  background-color: white;
-  box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.14), 0 2px 1px -1px rgba(0, 0, 0, 0.12), 0 1px 3px 0 rgba(0, 0, 0, 0.2);
-`;
-
-const StyledTable = styled(Table)`
-  ${({ searchView }) => searchView === 'table' && tableCss}
-`;
-
-export { ArrowUpOn, ArrowDownOn, ArrowUpOff, ArrowDownOff, StyledHeaderCell, StyledTable };
+export { ArrowUpOn, ArrowDownOn, ArrowUpOff, ArrowDownOff, StyledHeaderCell };
